gridland3/requirements.txt
requests>=2.31.0
click>=8.1.0
colorama>=0.4.6
scikit-learn>=1.0
joblib>=1.0
python-dotenv>=1.0.0
tabulate>=0.9.0
aiohttp>=3.8.0
PyYAML>=6.0
pyppeteer>=2.0.0
ffmpeg-python>=0.2.0
python-vlc>=3.0.20123
Flask>=2.0.0
Flask-WTF>=1.0.0
flask-talisman>=1.0.0
gunicorn>=20.0.0
pytest>=8.0.0
pytest-mock>=3.12.0